move shared code to a better location
