+++ apps/designer/app/shared/pages/index.ts
@@ -1,0 +1,1 @@
+export * from "./types";
