+++ apps/designer/app/shared/pages/types.ts
@@ -1,0 +1,10 @@
+import { type Page } from "@webstudio-is/project";
+import { FetcherData } from "~/shared/form-utils";
+
+export type CreatePageData = FetcherData<{ page: Page }>;
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type EditPageData = FetcherData<{}>;
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type DeletePageData = FetcherData<{}>;
